+++ RELEASENOTES.md
@@ -54,6 +54,9 @@
     *   Remove `Renderer[]` parameter from `LoadControl.onTracksSelected()` as
         `DefaultLoadControl` implementation can retrieve the stream types from
         `ExoTrackSelection[]`.
+    *   Add a setter to `SntpClient` to set the max elapsed time since the last
+        update after which the client is re-initialized
+        ([#1794](https://github.com/androidx/media/pull/1794)).
     *   Deprecated `DefaultLoadControl.calculateTargetBufferBytes(Renderer[],
         ExoTrackSelection[])` and marked method as final to prevent overrides.
         The new